型をprismaで生成されたinterfaceに置き換え
