+++ front/react_app/src/pages/cycle/Cycle.tsx
@@ -7,7 +7,7 @@ import { useParams } from "react-router-dom";
 import Layout from "../../components/Layout";
 import { refreshCycle, setCycle } from "../../redux/cycleSlice";
 import { RootState } from "../../redux/store";
-import { EditCycleIfc } from "../../utils/cycle.interface";
+import { EditCycleIfc } from "../../utils/interface";
 import AboutCycle from "./AboutCycle";
 import PlanSidebar from "./plan/PlanSidebar";
