+++ front/react_app/src/utils/interface.ts
@@ -1,0 +1,7 @@
+export interface EditCycleIfc {
+  name: string;
+  about: string | null;
+  goal: string;
+  watchFromAnyone: boolean;
+  suspend: boolean;
+}
