--- front/react_app/src/utils/plan.interface.ts
@@ -1,6 +1,0 @@
-export interface Plan {
-  cycleId: number;
-  round: number;
-  goalInRound: string | null
-  complete: boolean
-}
